# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Multi-Store Inventory Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List stores
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [stores]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /stores/{id}:
    get:
      operationId: getStore
      tags: [stores]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
    patch:
      operationId: updateStore
      tags: [stores]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"

  # CATEGORIES
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [raw_material, semi_finished, finished, packaging]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SUPPLIERS
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierList"
    post:
      operationId: createSupplier
      tags: [suppliers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"

  # CUSTOMERS
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [paid, credit, partially_paid, overdue]
        - name: customerId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleList"
    post:
      operationId: createSale
      tags: [sales]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  # PAYMENTS
...
components:
  schemas:
...
    Store:
      type: object
      required: [id, name, type, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [main_raw_material, semi_finished, final_production, main_finished]
        location:
          type: ["string", "null"]
        isActive:
          type: boolean
        managerId:
          type: ["integer", "null"]
    StoreInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        location:
          type: ["string", "null"]
        managerId:
          type: ["integer", "null"]
    StoreUpdate:
      type: object
      properties:
        name:
          type: string
        location:
          type: ["string", "null"]
        isActive:
          type: boolean
        managerId:
          type: ["integer", "null"]

    # CATEGORY
...
    Product:
      type: object
      required: [id, name, type, unit, reorderLevel, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: ["string", "null"]
        type:
          type: string
          enum: [raw_material, semi_finished, finished, packaging]
        unit:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        reorderLevel:
          type: number
        unitCost:
          type: ["number", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    ProductList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProductInput:
      type: object
      required: [name, type, unit]
      properties:
        name:
          type: string
        sku:
          type: ["string", "null"]
        type:
          type: string
        unit:
          type: string
        categoryId:
          type: ["integer", "null"]
        reorderLevel:
          type: number
        unitCost:
          type: ["number", "null"]
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        sku:
          type: ["string", "null"]
        type:
          type: string
        unit:
          type: string
        categoryId:
          type: ["integer", "null"]
        reorderLevel:
          type: number
        unitCost:
          type: ["number", "null"]
        isActive:
          type: boolean

    # SUPPLIER
    Supplier:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        taxNumber:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    SupplierList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Supplier"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    SupplierInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        contactPerson:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        taxNumber:
          type: ["string", "null"]
    SupplierUpdate:
      type: object
      properties:
        name:
          type: string
        contactPerson:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        taxNumber:
          type: ["string", "null"]
        isActive:
          type: boolean

    # CUSTOMER
...
    SaleItem:
      type: object
      required: [productId, quantity, unit, unitPrice]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
        discount:
          type: number
        totalPrice:
          type: number
    Sale:
      type: object
      required: [id, invoiceNumber, saleDate, customerId, paymentType, status, totalAmount, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        saleDate:
          type: string
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        fsNumber:
          type: ["string", "null"]
        paymentType:
          type: string
          enum: [cash, credit]
        paymentMethod:
          type: ["string", "null"]
        status:
          type: string
          enum: [paid, credit, partially_paid, overdue]
        subtotal:
          type: number
        vatApplicable:
          type: boolean
        vatAmount:
          type: number
        withholdingAmount:
          type: number
        discountAmount:
          type: number
        totalAmount:
          type: number
        paidAmount:
          type: number
        balanceDue:
          type: number
        dueDate:
          type: ["string", "null"]
        salespersonId:
          type: ["integer", "null"]
        salespersonName:
          type: ["string", "null"]
        storeId:
          type: ["integer", "null"]
        remarks:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        createdAt:
          type: string
    SaleList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    SaleInput:
      type: object
      required: [customerId, saleDate, paymentType, items]
      properties:
        customerId:
          type: integer
        saleDate:
          type: string
        fsNumber:
          type: ["string", "null"]
        paymentType:
          type: string
        paymentMethod:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        vatApplicable:
          type: boolean
        dueDate:
          type: ["string", "null"]
        salespersonId:
          type: ["integer", "null"]
        storeId:
          type: ["integer", "null"]
        remarks:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"

    # PAYMENT
...